+++ man-roxygen/Dbms.R
@@ -11,4 +11,5 @@
 #' - "sqlite extended" for SQLite with extended types (DATE and DATETIME)
 #' - "spark" for Spark
 #' - "snowflake" for Snowflake
+#' - "iris" for InterSystems IRIS
 #'