+++ NEWS.md
@@ -1,3 +1,22 @@
+DatabaseConnector 6.4.1
+=======================
+
+Bugfixes:
+
+- Remove direct call to `bit64` S3 method to avoid issues in the future.
+
+- Fixed error when calling `getTableNames()` on a `DuckDB` connection.
+
+
+DatabaseConnector 6.4.0
+=======================
+
+Changes:
+
+- Adding support for InterSystems IRIS.
+
+
+
 DatabaseConnector 6.3.3
 =======================
 
@@ -6,6 +25,10 @@ Changes:
 - Updated Snowflake driver to 3.16.1.
 
 - Updated Databricks driver to 2.6.36.
+
+- Updated BigQuery driver to 1.6.2.
+
+- Using `INSERT` with multiple values in `insertTable()` for DataBricks for faster inserts.
 
 
 DatabaseConnector 6.3.2